+++ packages/client/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.0.18-beta
 - Start fetching data when `active$` is not defined
-- Add actions graph client
+- Add actions graph, actions list client
 - Fix `destroy` function entity client
 - Use pipe instead of let operator
 - Use `DateRangeMap` for `initialDateRange`